flask/requirements.txt
Flask
Flask-Cors
tensorflow
numpy
pillow
pandas
seaborn
matplotlib
scikit-learn